% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aai.R
\name{aai}
\alias{aai}
\title{aai}
\usage{
aai(pattern = ".faa", reference = "all", path = ".",
  outdir = "aai_output", proc = 2)
}
\arguments{
\item{pattern}{a pattern (like '.faa') for recognizing the genome files.}

\item{reference}{is a genome name used to compare one vs. the rest. If set to 'all' (default), all vs. all is computed.}

\item{path}{is the full path to where genome annotations in fasta format are placed.}

\item{outdir}{a name for the output directory that will be created for results.}

\item{proc}{is the number of threads for parallel computing.}
}
\description{
Calculates AAI (Average Amino acid Identity) values for all genomes present in the folder.
}
\examples{
aai(path='./prodigal_output',pattern='.faa',reference='all',outdir='aai_result',proc=2)
}
\keyword{AAI}
\keyword{species}

