% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agios.R
\name{agios}
\alias{agios}
\title{agios}
\usage{
agios(pattern = ".ffn", reference = "all", path = ".",
  outdir = "agios_output", proc = 2)
}
\arguments{
\item{pattern}{a pattern (like '.ffn') for recognizing the genome annotation files in nucleotides.}

\item{reference}{is a genome name used to compare one vs. the rest. If set to 'all' (default), all vs. all is computed.}

\item{path}{is the full path to where genome annotations in fasta format are placed.}

\item{outdir}{a name for the output directory that will be created for results.}

\item{proc}{is the number of threads for parallel computing.}
}
\description{
Calculates AGIOS (Average Genome Identity of Orthologous Sequences) values for all genomes present in the folder.
}
\examples{
agios(path='./prodigal_output',pattern='.ffn',reference='all',outdir='agios_result',proc=2)
}
\keyword{AGIOS}
\keyword{species}

