% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anib.R
\name{anib}
\alias{anib}
\title{anib}
\usage{
anib(pattern = ".fna", reference = "all", path = ".",
  outdir = "anib_output", proc = 2)
}
\arguments{
\item{pattern}{a pattern (like '.fna') for recognizing the genome files.}

\item{reference}{is a genome name used to compare one vs. the rest. If set to 'all' (default), all vs. all is computed.}

\item{path}{is the full path to where genomes assemblies in fasta format are placed.}

\item{outdir}{a name for the output directory that will be created for results.}

\item{proc}{is the number of threads for parallel computing.}
}
\description{
Calculates ANIb (Average Nucleotide Identity using Blast) values for all genomes present in a directory.
}
\examples{
anib(path='.',pattern='.fna',reference='all',outdir='anib_result',proc=2)
}
\keyword{ANI}
\keyword{blastn}
\keyword{species}

