% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprot.R, R/mgene.R
\name{aprot}
\alias{aprot}
\title{aprot}
\usage{
aprot(pattern = ".faa", path = ".", outdir = "mgene_output", marker,
  proc = 2)

aprot(pattern = ".faa", path = ".", outdir = "mgene_output", marker,
  proc = 2)
}
\arguments{
\item{pattern}{a pattern (like '.faa') for recognizing the annotation files.}

\item{path}{is the full path to where genome annotations in fasta format are placed.}

\item{outdir}{a name for the output directory that will be created for results.}

\item{marker}{is the multifasta file in amino acids for the desired marker gene.}

\item{proc}{is the number of threads used for protein searches.}

\item{path}{is the full path to where genome annotations in fasta format are placed.}

\item{pattern}{is a pattern (like '.faa') for recognizing the annotation files.}

\item{outdir}{is a name for the output directory that will be created for results.}

\item{proc}{is the number of threads used for protein searches.}
}
\description{
Identifies and extracts a set of 37 archaeal universal proteins used as phylogenetic markers.

Identifies and extracts a desired gene used as taxonomic/phylogenetic marker.
}
\examples{
aprot(path='./prodigal_output',pattern='.faa',outdir='aprot_output',proc=2)
mgene(path='./prodigal_output',pattern='.faa',marker='gyrB.faa',outdir='mgene_output',proc=2)
}
\keyword{archaeal}
\keyword{gene}
\keyword{marker}
\keyword{markers}
\keyword{proteins}

