% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodigal.R
\name{prodigal}
\alias{prodigal}
\title{prodigal}
\usage{
prodigal(path = ".", pattern = ".fasta", outdir = "prodigal_output",
  proc = 2)
}
\arguments{
\item{path}{is the full path to where genomes in fasta format are placed.}

\item{pattern}{a pattern (like '.fasta') for recognizing the genome files.}

\item{outdir}{a name for the output directory that will be created for results.}
}
\description{
Predicts protein coding genes from the genome using Prodigal.
}
\examples{
rrna(path='.',pattern='.fasta',outdir='prodigal_output')
}
\keyword{Prodigal}
\keyword{gene}
\keyword{prediction}

