% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrna.R
\name{rrna}
\alias{rrna}
\title{rrna}
\usage{
rrna(path, pattern, outdir = "rrna_out", subunit = "16S",
  kingdom = "bacteria", align = TRUE)
}
\arguments{
\item{path}{is the full path to where genomes in fasta format are placed.}

\item{pattern}{a pattern (like '.fasta') for recognizing the genome files.}

\item{outdir}{a name for the output directory that will be created for results.}

\item{subunit}{takes '16S' (default), '23S' or 'both'.}

\item{kingdom}{takes 'bacteria' (default) or 'archaea'.}

\item{align}{takes a logical (default, TRUE) indicating if sequence alignment is performed.}
}
\description{
Outputs a individual rRNA gene sequences and a multiple sequence alignment.
}
\examples{
rrna(path='.',pattern='.fasta',outdir='rrna_output',subunit='16S',kingdom='bacteria')
}
\keyword{16S}
\keyword{23S}
\keyword{rRNA}

