% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uprot.R
\name{uprot}
\alias{uprot}
\title{uprot}
\usage{
uprot(pattern = ".faa", path = ".", outdir = "uprot_output", proc = 2)
}
\arguments{
\item{pattern}{a pattern (like '.faa') for recognizing the annotation files.}

\item{path}{is the full path to where genome annotations in fasta format are placed.}

\item{outdir}{a name for the output directory that will be created for results.}

\item{proc}{is the number of threads used for protein searches.}
}
\description{
Identifies and extracts a set of 40 prokaryotic universal proteins used as phylogenetic markers.
}
\examples{
uprot(path='./prodigal_output',pattern='.faa',outdir='uprot_output',proc=2)
}
\keyword{markers}
\keyword{proteins}
\keyword{ribosomal}
\keyword{universal}

